fix compile/linkage errors on Mac
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2024 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/platform/unix/RexxQueues.cpp
@@ -61,7 +61,7 @@
 RexxMethod0(REXXOBJECT, rexx_query_queue)
 {
    REXXOBJECT queue_name;              /* current queue name                */
-   unsigned int count = 0;             /* count of lines                    */
+   size_t count = 0;                   /* count of lines                    */
    APIRET rc;                          /* queue query return code           */
 
                                        /* get the queue name                */
